split out cert verify calls
