+++ library/common/jni/android_jni_interface.cc
@@ -1,6 +1,6 @@
 #include <ares.h>
 
-#include "library/common/jni/android_jni_utility.h"
+#include "library/common/jni/android_network_utility.h"
 #include "library/common/jni/import/jni_import.h"
 #include "library/common/jni/jni_support.h"
 #include "library/common/jni/jni_utility.h"
